Revert "Move `useLogVersionInfo` to the SDK npm package"
This reverts commit 8c71459a3374ea095533df4acdc5bd304e9f57c8.
