+++ enterprise/frontend/src/embedding-sdk/hooks/private/use-log-version-info/index.ts
