+++ enterprise/frontend/src/embedding-sdk/lib/get-embedding-sdk-package-build-data.ts
