--- enterprise/frontend/src/embedding-sdk/sdk-package/store/selectors.ts
@@ -1,4 +1,0 @@
-import type { SdkStoreState } from "embedding-sdk/store/types";
-
-export const getSdkBundleVersion = (state: SdkStoreState) =>
-  state.settings.values.version.tag;
