--- enterprise/frontend/src/embedding-sdk/sdk-shared/lib/is-invalid-metabase-version.ts
@@ -1,8 +1,0 @@
-// They are mostly used for local development
-export const isInvalidMetabaseVersion = (mbVersion: string) => {
-  return (
-    mbVersion === "vLOCAL_DEV" ||
-    mbVersion === "vUNKNOWN" ||
-    mbVersion.endsWith("-SNAPSHOT")
-  );
-};
