--- enterprise/frontend/src/embedding-sdk/sdk-shared/lib/is-sdk-package-compatible-with-sdk-bundle.ts
@@ -1,9 +1,0 @@
-import { getMajorVersion } from "embedding-sdk/sdk-shared/lib/get-major-version";
-
-export const isSdkPackageCompatibleWithSdkBundle = ({
-  sdkPackageVersion,
-  sdkBundleVersion,
-}: {
-  sdkPackageVersion: string;
-  sdkBundleVersion: string;
-}) => getMajorVersion(sdkPackageVersion) === getMajorVersion(sdkBundleVersion);
